refactor: migrate branch resolvers from utils to services
- Move branch_arg.py: utils/ → services/
- Move pr_branch_resolver.py: utils/ → services/
- Update all import references in commands and services

These modules contain business logic and depend on services/clients,
so they belong in services/ not utils/.

Related: #1245
