+++ src/vibe3/commands/flow_manage.py
@@ -167,8 +167,8 @@ def update(
         output_format = "json"
     from vibe3.clients.git_client import GitClient
     from vibe3.config.orchestra_settings import load_orchestra_config
+    from vibe3.services.branch_arg import resolve_branch_arg
     from vibe3.services.convention_resolver import ConventionResolver
-    from vibe3.utils.branch_arg import resolve_branch_arg
     from vibe3.utils.issue_ref import try_parse_issue_number
 
     # Early handling for issue number: resolve to canonical branch
@@ -442,7 +442,7 @@ def restore_flow(
         typer.echo("Error: Branch is required for flow restore", err=True)
         raise typer.Exit(1)
 
-    from vibe3.utils.branch_arg import resolve_branch_arg
+    from vibe3.services.branch_arg import resolve_branch_arg
 
     target_branch = resolve_branch_arg(branch)
