+++ src/vibe3/commands/handoff_read.py
@@ -18,9 +18,9 @@ from vibe3.exceptions import SystemError, UserError
 from vibe3.services.flow_service import FlowService
 from vibe3.services.handoff_status_service import HandoffStatusService
 from vibe3.services.issue_branch_resolver import resolve_issue_branch_input
+from vibe3.services.pr_branch_resolver import resolve_command_branch
 from vibe3.ui.console import console
 from vibe3.ui.handoff_ui import render_handoff_detail
-from vibe3.utils.pr_branch_resolver import resolve_command_branch
 
 
 def _format_relative_time(timestamp: datetime, now: datetime | None = None) -> str:
