+++ src/vibe3/commands/handoff_write.py
@@ -7,10 +7,10 @@ from loguru import logger
 
 from vibe3.commands.common import trace_scope
 from vibe3.models.verdict_types import VerdictValue
+from vibe3.services.branch_arg import resolve_branch_arg
 from vibe3.services.handoff_service import HandoffService
 from vibe3.services.verdict_service import VerdictService
 from vibe3.ui.console import console
-from vibe3.utils.branch_arg import resolve_branch_arg
 
 
 def _record_handoff_reference(
