+++ src/vibe3/commands/plan.py
@@ -16,8 +16,8 @@ from vibe3.roles.plan import (
     execute_spec_plan_sync,
     resolve_spec_plan_input,
 )
+from vibe3.services.branch_arg import resolve_branch_arg
 from vibe3.services.flow_service import FlowService
-from vibe3.utils.branch_arg import resolve_branch_arg
 from vibe3.utils.handoff_resolution import resolve_handoff_target
 from vibe3.utils.trace import enable_trace
