+++ src/vibe3/commands/review.py
@@ -25,8 +25,8 @@ from vibe3.roles.review import (
     execute_manual_review_sync,
     validate_review_prerequisites,
 )
+from vibe3.services.branch_arg import resolve_branch_arg
 from vibe3.services.flow_service import FlowService
-from vibe3.utils.branch_arg import resolve_branch_arg
 from vibe3.utils.trace import enable_trace
 
 app = typer.Typer(
